refactor(apps/effect-api-server): split execution from definition
https.ts contains the building of the
http layer while index.ts contains the
node runtime execution
